Refactor: Consolidate SignatureAI page and remove duplicate component
